requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
jinja2
python-multipart
python-dotenv
httpx
passlib[bcrypt]
aiofiles